services/friendship-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
python-jose[cryptography]
passlib[bcrypt]
prometheus-client
prometheus-fastapi-instrumentator
aio-pika
